Move sleep method from TTS utils to BR utils
